Add a separate tsconfig for typecheck test
The reason to do this is because if we run `tsc --noEmit` (typical type check command) via yarn, it throws an error because our tsconfig.json has 'emitDeclarationOnly' set to true. This conflicts with the noEmit flag. In order to circumvent this (because we want to be able to just typecheck occasionally), I've added an extra config which extends tsconfig.