Bundling with correct type defs (#37)
* Temporarily fix type-def issue in parcel build

* Remove typecheck from test

* Add a separate tsconfig for typecheck test

The reason to do this is because if we run `tsc --noEmit` (typical type check command) via yarn, it throws an error because our tsconfig.json has 'emitDeclarationOnly' set to true. This conflicts with the noEmit flag. In order to circumvent this (because we want to be able to just typecheck occasionally), I've added an extra config which extends tsconfig.

* Remove keys from options - fix issue #40 (#42)

* Fix #40:

Remove query and data keys from options before passing it to asyncFetchWrapper

* Change beforeEach to beforeAll (#43)